fix: remove unnecessary cast of boolean to int in Cfg Save()

+++ src/core/configuration.cpp
@@ -196,7 +196,7 @@ ConfigResult Configuration::Save()
                 { "id", static_cast<int>(tasksViewColumn.TaskViewColumnId) },
                 { "width", tasksViewColumn.Width },
                 { "type", static_cast<int>(tasksViewColumn.Type) },
-                { "selected", static_cast<int>(tasksViewColumn.Selected) }
+                { "selected", tasksViewColumn.Selected }
             }
         );
         // clang-format on
